pyproject.toml
[project]
name = "ku-eater-embagent"
version = "0.1.0"
description = "KU Eater's embedding agent & recommendation engine"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "psycopg[binary,pool]~=3.2",        # pgsql
    "pandas~=2.2",                      # dataframe
    "torch~=2.5",                       # pytorch
    "transformers~=4.48",               # transformers
    "sentence-transformers[onnx]~=4.0", # sentence-transformers
    "grpcio==1.71.0",                   # gRPC
    "python-dotenv~=1.1"
]

[dependency-groups]
data = [
    "uuid7~=0.1",
    "sqlfluff~=3.3"
]
dev = [
    "ruff",
    "grpcio-tools==1.71.0"
]

[tool.uv.sources]
torch = [
    { index = "torch-linux", marker = "platform_system == 'Linux'" }
]

[[tool.uv.index]]
name = "torch-linux"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[tool.ruff]
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]

line-length = 88
indent-width = 4

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.sqlfluff.core]
dialect = "postgres"
encoding = "utf-8"
large_file_skip_char_limit = 0
large_file_skip_byte_limit = 0